HTML and CSS files created
Two files were created with the name index.html and styles.css and linked to each other. In index.html, we have added the header section, given the title and heading, and used Google fonts to provide the site with a unique and stylish look. Also, In styles.css, we did set styling for different devices and header sections.